+++ plugins/bulk-import-backend/src/helpers/catalogInfoGenerator.ts
@@ -16,6 +16,7 @@
 import { PluginEndpointDiscovery } from '@backstage/backend-common';
 import { AuthService, DiscoveryService } from '@backstage/backend-plugin-api';
 import { CatalogApi } from '@backstage/catalog-client';
+import { LocationEntity } from '@backstage/catalog-model';
 import type { Config } from '@backstage/config';
 
 import gitUrlParse from 'git-url-parse';
@@ -134,12 +135,13 @@ ${jsYaml.dump(generatedEntity.entity)}`,
   async listCatalogUrlLocationsById(
     config: Config,
     search?: string,
-    _pageNumber: number = DefaultPageNumber,
-    _pageSize: number = DefaultPageSize,
+    pageNumber: number = DefaultPageNumber,
+    pageSize: number = DefaultPageSize,
   ): Promise<{ id?: string; target: string }[]> {
     const result = await Promise.all([
       this.listCatalogUrlLocationsFromConfig(config, search),
       this.listCatalogUrlLocationsByIdFromLocationsEndpoint(search),
+      this.listCatalogUrlLocationEntitiesById(search, pageNumber, pageSize),
     ]);
     return result.flat();
   }
@@ -196,6 +198,39 @@ ${jsYaml.dump(generatedEntity.entity)}`,
     return this.filterLocations(res, search);
   }
 
+  async listCatalogUrlLocationEntitiesById(
+    search?: string,
+    _pageNumber: number = DefaultPageNumber,
+    _pageSize: number = DefaultPageSize,
+  ): Promise<{ id?: string; target: string }[]> {
+    const result = await this.catalogApi.getEntities(
+      {
+        filter: {
+          kind: 'Location',
+        },
+        // There is no query parameter to find entities with target URLs containing a string.
+        // The existing filter does an exact matching. That's why we are retrieving this hard-coded high number of Locations.
+        limit: 1000,
+        offset: 0,
+      },
+      {
+        token: await getTokenForPlugin(this.auth, 'catalog'),
+      },
+    );
+    const locations = (result?.items ?? []) as LocationEntity[];
+    const res = locations
+      .filter(
+        location => location.spec?.target && location.spec?.type === 'url',
+      )
+      .map(location => {
+        return {
+          id: location.metadata.uid,
+          target: location.spec.target!,
+        };
+      });
+    return this.filterLocations(res, search);
+  }
+
   private filterLocations(
     res: { id: string | undefined; target: string }[],
     search: string | undefined,
@@ -232,6 +267,75 @@ ${jsYaml.dump(generatedEntity.entity)}`,
 
   async deleteCatalogLocationEntityById(locationUid: string): Promise<void> {
     await this.catalogApi.removeEntityByUid(locationUid, {
+      token: await getTokenForPlugin(this.auth, 'catalog'),
+    });
+  }
+
+  async findLocationEntitiesByRepoUrl(
+    config: Config,
+    repoUrl: string,
+    defaultBranch?: string,
+  ) {
+    return this.findLocationEntitiesByTargetUrl(
+      this.getCatalogUrl(config, repoUrl, defaultBranch),
+    );
+  }
+
+  async findLocationEntitiesByTargetUrl(targetUrl: string, limit?: number) {
+    return this.catalogApi
+      .queryEntities(
+        {
+          filter: [
+            { kind: 'Location', 'spec.type': 'url', 'spec.target': targetUrl },
+          ],
+          fields: ['metadata.namespace', 'metadata.name', 'metadata.uid'],
+          limit,
+        },
+        {
+          token: await getTokenForPlugin(this.auth, 'catalog'),
+        },
+      )
+      .then(resp => resp.items);
+  }
+
+  async refreshLocationByRepoUrl(
+    config: Config,
+    repoUrl: string,
+    defaultBranch?: string,
+  ) {
+    const promises: Promise<void>[] = [];
+    this.findLocationEntitiesByRepoUrl(config, repoUrl, defaultBranch).then(
+      entities => {
+        const nbEntities = entities.length;
+        if (nbEntities === 0) {
+          this.logger.debug(`No Location Entity found for repo: ${repoUrl}`);
+          return;
+        }
+        this.logger.debug(
+          `Refreshing ${nbEntities} Location(s) for repo: ${repoUrl}`,
+        );
+        entities.forEach(ent =>
+          promises.push(
+            this.refreshEntity(
+              'location',
+              ent.metadata.name,
+              ent.metadata.namespace,
+            ),
+          ),
+        );
+      },
+    );
+    await Promise.all(promises);
+  }
+
+  async refreshEntity(
+    kind: string,
+    name: string,
+    namespace: string = 'default',
+  ) {
+    const entityRef = `${kind}:${namespace}/${name}`;
+    this.logger.debug(`Refreshing entityRef: ${entityRef}`);
+    await this.catalogApi.refreshEntity(entityRef, {
       token: await getTokenForPlugin(this.auth, 'catalog'),
     });
   }