+++ plugins/bulk-import-backend/src/service/router.test.ts
@@ -21,7 +21,12 @@ import {
   BackstagePrincipalTypes,
   CacheService,
 } from '@backstage/backend-plugin-api';
-import { CatalogClient } from '@backstage/catalog-client';
+import {
+  CatalogClient,
+  CatalogRequestOptions,
+  QueryEntitiesRequest,
+  QueryEntitiesResponse,
+} from '@backstage/catalog-client';
 import { ConfigReader } from '@backstage/config';
 import {
   AuthorizeResult,
@@ -169,7 +174,9 @@ describe('createRouter', () => {
       getEntitiesByRefs: mockGetEntitiesByRefs,
       validateEntity: mockValidateEntity,
       addLocation: mockAddLocation,
+      getEntities: jest.fn(),
       queryEntities: jest.fn(),
+      refreshEntity: jest.fn(),
     } as unknown as CatalogClient;
     const voidLogger = getVoidLogger();
     mockCatalogInfoGenerator = new CatalogInfoGenerator(
@@ -960,6 +967,15 @@ describe('createRouter', () => {
         )
         .mockResolvedValue(fromLocationsEndpoint);
       jest
+        .spyOn(
+          mockGithubApiService,
+          'filterLocationsAccessibleFromIntegrations',
+        )
+        .mockImplementation((locationUrls: string[]) => {
+          // filter returning the same input
+          return Promise.resolve(locationUrls);
+        });
+      jest
         .spyOn(mockGithubApiService, 'getRepositoryFromIntegrations')
         .mockImplementation(repoUrl => {
           let defaultBranch: string | undefined;
@@ -1028,6 +1044,30 @@ describe('createRouter', () => {
             ].includes(input.repoUrl),
           );
         });
+
+      mockCatalogClient.queryEntities = jest
+        .fn()
+        .mockImplementation(
+          async (
+            _request?: QueryEntitiesRequest,
+            _options?: CatalogRequestOptions,
+          ): Promise<QueryEntitiesResponse> => {
+            return {
+              items: [
+                {
+                  apiVersion: 'backstage.io/v1alpha1',
+                  kind: 'Location',
+                  metadata: {
+                    name: `generated-from-tests-${Math.floor(Math.random() * 100 + 1)}`,
+                    namespace: 'default',
+                  },
+                },
+              ],
+              totalItems: 1,
+              pageInfo: {},
+            };
+          },
+        );
 
       const response = await request(app).get('/imports');
       expect(response.status).toEqual(200);
@@ -1132,6 +1172,8 @@ describe('createRouter', () => {
           status: null,
         },
       ]);
+      // Location entity refresh triggered twice (on each 'ADDED' repo)
+      expect(mockCatalogClient.refreshEntity).toHaveBeenCalledTimes(2);
     });
   });
 
@@ -1200,6 +1242,29 @@ describe('createRouter', () => {
               );
           }
         });
+      mockCatalogClient.queryEntities = jest
+        .fn()
+        .mockImplementation(
+          async (
+            _request?: QueryEntitiesRequest,
+            _options?: CatalogRequestOptions,
+          ): Promise<QueryEntitiesResponse> => {
+            return {
+              items: [
+                {
+                  apiVersion: 'backstage.io/v1alpha1',
+                  kind: 'Location',
+                  metadata: {
+                    name: `generated-from-tests-${Math.floor(Math.random() * 100 + 1)}`,
+                    namespace: 'default',
+                  },
+                },
+              ],
+              totalItems: 1,
+              pageInfo: {},
+            };
+          },
+        );
 
       const response = await request(app)
         .post('/imports')
@@ -1275,6 +1340,8 @@ spec:
           status: 'ADDED',
         },
       ]);
+      // Location entity refresh triggered (on each 'ADDED' repo)
+      expect(mockCatalogClient.refreshEntity).toHaveBeenCalledTimes(1);
     });
   });
 
@@ -1301,7 +1368,7 @@ spec:
               ],
             };
           }
-          return { items: [] };
+          return { totalItems: 0, items: [] };
         },
       );