+++ plugins/bulk-import/src/utils/repository-utils.tsx
@@ -73,6 +73,7 @@ export const getComparator = (
 
 export const defaultCatalogInfoYaml = (
   componentName: string,
+  repoName: string,
   orgName: string,
   owner: string,
 ) => ({
@@ -80,10 +81,26 @@ export const defaultCatalogInfoYaml = (
   kind: 'Component',
   metadata: {
     name: componentName,
-    annotations: { 'github.com/project-slug': `${orgName}/${componentName}` },
+    annotations: { 'github.com/project-slug': `${orgName}/${repoName}` },
   },
   spec: { type: 'other', lifecycle: 'unknown', owner },
 });
+
+export const componentNameRegex =
+  /^([a-zA-Z0-9]+[-_.])*[a-zA-Z0-9]+$|^[a-zA-Z0-9]{1,63}$/;
+
+export const cleanComponentName = (input: string) => {
+  // Remove leading and trailing dots, underscores, dollar or hyphens
+  const cleanedStr = input.replace(/^[$._-]+|[$._-]+$/g, '');
+
+  if (componentNameRegex.test(input)) {
+    return input;
+  }
+  if (componentNameRegex.test(cleanedStr)) {
+    return cleanedStr;
+  }
+  return 'my-component';
+};
 
 export const getPRTemplate = (
   componentName: string,
@@ -96,13 +113,14 @@ export const getPRTemplate = (
   const importJobUrl = repositoryUrl
     ? `${baseUrl}/bulk-import/repositories?repository=${repositoryUrl}&defaultBranch=${defaultBranch}`
     : `${baseUrl}/bulk-import/repositories`;
+  const name = cleanComponentName(componentName);
   return {
-    componentName,
+    componentName: name,
     entityOwner,
     prTitle: 'Add catalog-info.yaml config file',
     prDescription: `This pull request adds a **Backstage entity metadata file**\nto this repository so that the component can\nbe added to the [software catalog](${baseUrl}/catalog).\nAfter this pull request is merged, the component will become available.\nFor more information, read an [overview of the Backstage software catalog](https://backstage.io/docs/features/software-catalog/).\nView the import job in your app [here](${importJobUrl}).`,
     useCodeOwnersFile: false,
-    yaml: defaultCatalogInfoYaml(componentName, orgName, entityOwner),
+    yaml: defaultCatalogInfoYaml(name, componentName, orgName, entityOwner),
   };
 };
 
@@ -399,18 +417,16 @@ export const getApi = (
   backendUrl: string,
   page: number,
   size: number,
-  options: APITypes,
+  searchString: string,
+  options?: APITypes,
 ) => {
-  if (options.fetchRepositories) {
-    return `${backendUrl}/api/bulk-import/repositories?pagePerIntegration=${page}&sizePerIntegration=${size}`;
+  if (options?.fetchOrganizations) {
+    return `${backendUrl}/api/bulk-import/organizations?pagePerIntegration=${page}&sizePerIntegration=${size}&search=${searchString}`;
   }
-  if (options.orgName) {
-    return `${backendUrl}/api/bulk-import/organizations/${options.orgName}/repositories?pagePerIntegration=${page}&sizePerIntegration=${size}`;
+  if (options?.orgName) {
+    return `${backendUrl}/api/bulk-import/organizations/${options.orgName}/repositories?pagePerIntegration=${page}&sizePerIntegration=${size}&search=${searchString}`;
   }
-  if (options.fetchOrganizations) {
-    return `${backendUrl}/api/bulk-import/organizations?pagePerIntegration=${page}&sizePerIntegration=${size}`;
-  }
-  return '';
+  return `${backendUrl}/api/bulk-import/repositories?pagePerIntegration=${page}&sizePerIntegration=${size}&search=${searchString}`;
 };
 
 export const getCustomisedErrorMessage = (
@@ -493,31 +509,46 @@ export const calculateLastUpdated = (dateString: string) => {
 export const evaluatePRTemplate = (
   repositoryStatus: ImportJobStatus,
 ): { pullReqPreview: PullRequestPreview; isInvalidEntity: boolean } => {
-  const entity = jsyaml.loadAll(
-    repositoryStatus.github.pullRequest.catalogInfoContent,
-  )[0] as Entity;
-  const isInvalid =
-    !entity?.metadata?.name || !entity?.apiVersion || !entity?.kind;
-  return {
-    pullReqPreview: {
-      pullRequestUrl: repositoryStatus.github.pullRequest.url,
-      prTitle: repositoryStatus.github.pullRequest.title,
-      prDescription: repositoryStatus.github.pullRequest.body,
-      prAnnotations: convertKeyValuePairsToString(
-        entity?.metadata?.annotations,
-      ),
-      prLabels: convertKeyValuePairsToString(entity?.metadata?.labels),
-      prSpec: convertKeyValuePairsToString(
-        entity?.spec as Record<string, string>,
-      ),
-      componentName: entity?.metadata?.name,
-      entityOwner: entity?.spec?.owner as string,
-      useCodeOwnersFile: !entity?.spec?.owner,
-      yaml: entity,
-    },
-    isInvalidEntity: isInvalid,
-  };
+  try {
+    const entity = jsyaml.loadAll(
+      repositoryStatus.github.pullRequest.catalogInfoContent,
+    )[0] as Entity;
+    const isInvalid =
+      !entity?.metadata?.name || !entity?.apiVersion || !entity?.kind;
+    return {
+      pullReqPreview: {
+        pullRequestUrl: repositoryStatus.github.pullRequest.url,
+        prTitle: repositoryStatus.github.pullRequest.title,
+        prDescription: repositoryStatus.github.pullRequest.body,
+        prAnnotations: convertKeyValuePairsToString(
+          entity?.metadata?.annotations,
+        ),
+        prLabels: convertKeyValuePairsToString(entity?.metadata?.labels),
+        prSpec: convertKeyValuePairsToString(
+          entity?.spec as Record<string, string>,
+        ),
+        componentName: entity?.metadata?.name,
+        entityOwner: entity?.spec?.owner as string,
+        useCodeOwnersFile: !entity?.spec?.owner,
+        yaml: entity,
+      },
+      isInvalidEntity: isInvalid,
+    };
+  } catch (e) {
+    return {
+      pullReqPreview: {
+        pullRequestUrl: repositoryStatus.github.pullRequest.url,
+        prTitle: repositoryStatus.github.pullRequest.title,
+        prDescription: repositoryStatus.github.pullRequest.body,
+        prAnnotations: undefined,
+        prLabels: undefined,
+        prSpec: undefined,
+        componentName: undefined,
+        entityOwner: undefined,
+        useCodeOwnersFile: false,
+        yaml: {} as Entity,
+      },
+      isInvalidEntity: true,
+    };
+  }
 };
-
-export const componentNameRegex =
-  /^([a-zA-Z0-9]+[-_.])*[a-zA-Z0-9]+$|^[a-zA-Z0-9]{1,63}$/;