+++ plugins/orchestrator/CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Dependencies
 
+* **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.17.2
+* **@janus-idp/backstage-plugin-orchestrator-form-react:** upgraded to 1.0.10
+
+### Dependencies
+
 * **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.17.1
 * **@janus-idp/backstage-plugin-orchestrator-form-react:** upgraded to 1.0.9